# go/model_course.go
/*
 * ITU Super API
 *
 * Test server for ITU Information
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Course struct {

	Id int32 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	// The IDs of the teachers teaching this course.
	Teachers []int32 `json:"teachers,omitempty"`

	Coursework int32 `json:"coursework,omitempty"`
}