# go/model_course_rating.go
/*
 * ITU Super API
 *
 * Test server for ITU Information
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type CourseRating struct {

	// The student giving the rating
	StudentId int32 `json:"student_id,omitempty"`

	// The course getting rated
	CourseId int32 `json:"course_id,omitempty"`

	// The rating the student is giving
	Rating int32 `json:"rating,omitempty"`
}