# go/model_student.go
/*
 * ITU Super API
 *
 * Test server for ITU Information
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Student struct {

	Id int32 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	// IDs of courses taken by student
	Courses []int32 `json:"courses,omitempty"`

	Workload int32 `json:"workload,omitempty"`
}