# go/model_teacher.go
/*
 * ITU Super API
 *
 * Test server for ITU Information
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Teacher struct {

	Id int32 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	// IDs of courses taught by teacher
	CoursesTaught []int32 `json:"coursesTaught,omitempty"`
}